Cleanup PluginListItem
