+++ js/components/PluginListItem/Name.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import constants from '../../constants';
+
+export default styled.h2`
+  display: inline-block;
+  color: ${constants.textColor};
+  margin-top: 0;
+  margin-bottom: 0;
+  margin-right: 1em;
+`;
